python/requirements.txt
# 만세력 계산 API 의존성
# Python 3.11+

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Data Validation
pydantic>=2.5.0

# 만세력/음력 계산 (핵심 라이브러리)
lunar-python>=1.3.0

# 날짜/시간 처리
python-dateutil>=2.8.2
pytz>=2024.1

# 이미지 생성 (시각화)
Pillow>=10.2.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
